backend/requirements.txt
fastapi==0.115.12
SQLAlchemy==2.0.41
alembic==1.16.0
asyncpg==0.30.0
bcrypt==4.3.0
PyJWT==2.10.1